+++ examples/express/vite.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from 'vite';
+import { VitePluginNode } from 'vite-plugin-node';
+
+export default defineConfig({
+  plugins: [
+    ...VitePluginNode({
+      handler: 'express',
+      appPath: './app.ts'
+    })
+  ]
+})