+++ examples/koa/app.ts
@@ -3,7 +3,7 @@ import Koa from 'koa';
 const app = new Koa();
 
 app.use(async ctx => {
-  ctx.body = 'Change Me and Refresh to see HMR';
+  ctx.body = 'Change Me and Refresh to see HMR!';
 });
 
 if (process.env.NODE_ENV === 'production') {