+++ examples/nest/vite.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'vite';
+import { VitePluginNode } from 'vite-plugin-node';
+
+export default defineConfig({
+  plugins: [
+    ...VitePluginNode({
+      handler: 'nest',
+      appPath: './src/main.ts',
+      tsCompiler: 'swc'
+    })
+  ]
+})