+++ src/server/express.ts
@@ -1,0 +1,6 @@
+import { Application } from 'express';
+import { RequestHandler } from "..";
+
+export const ExpressHandler: RequestHandler<Application> = (app, req, res) => {
+  app(req, res)
+}